# api/openapi.yaml
openapi: 3.1.0
info:
  title: TOT API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    JSON API for the Next.js frontend (`frontend-tot`). Draft contract — see
    `docs/frontend/inventory.md` for the page inventory and open decisions.

    Conventions:
    - Base path `/api/v1`; frontend and API share one origin (reverse proxy), so the
      session cookie stays `HttpOnly; SameSite=Lax` and no CORS is needed.
    - JSON bodies and fields are `snake_case`; timestamps are RFC 3339.
    - Errors always use the `Error` schema with a stable machine `code`; user-facing
      text is localized on the frontend by `code`, `message` is for logs only.
    - Unauthenticated requests get `401`, never a redirect; malformed JSON gets `400 malformed_json`.
    - Mutating requests (POST/PUT/PATCH/DELETE) must come from an allowed origin
      (`Origin` / `Sec-Fetch-Site`), otherwise `403 csrf_rejected`.
    - Request bodies are `application/json` unless stated otherwise, otherwise `415 unsupported_media_type`.
    - HTML fields (`*_html`) are rendered and sanitized server-side.
    - Admin endpoints need role `author` or `admin`; admin-only ones say so. A course
      the author cannot access answers `404`; a missing permission on an accessible
      course answers `403 forbidden` (see `CourseAccess`).
servers:
  - url: /api/v1
...
security:
  - session: []
paths:
...
  /courses/{courseSlug}/cover:
    parameters:
      - $ref: '#/components/parameters/CourseSlug'
    get:
      tags: [catalog]
      summary: Get course cover
      operationId: getCourseCover
      security: []
      responses:
        '200':
          description: Uploaded image or generated SVG.
          content:
            image/*: { schema: { type: string, format: binary } }
            image/svg+xml: { schema: { type: string } }
        '302': { description: Redirect to an external cover URL. }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────────── lessons ─────────────────────────────
...
  /admin/courses:
    get:
      tags: [admin-courses]
      summary: List courses
      operationId: adminListCourses
      description: Admins get every course; authors get the courses they own or co-author.
      responses:
        '200':
          description: Courses.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminCourseRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin-courses]
      summary: Create course
      operationId: adminCreateCourse
      description: The creator becomes the owner. Draft unless `published` is true.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCourseInput' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCourse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/SlugTaken' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/courses/{courseId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [admin-courses]
      summary: Get course
      operationId: adminGetCourse
      responses:
        '200':
          description: Course with all lessons (drafts included); `404` without access.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCourseDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [admin-courses]
      summary: Update course
      operationId: adminUpdateCourse
      description: Full replace of editable fields (`published` and `cover_url` are kept when omitted). Needs `can_edit`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCourseInput' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCourse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/SlugTaken' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [admin-courses]
      summary: Delete course
      operationId: adminDeleteCourse
      description: Hard delete; cascades to lessons, quizzes, tasks, submissions and progress. Needs `can_delete`.
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/courses/{courseId}/published:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [admin-courses]
      summary: Set course published
      operationId: adminSetCoursePublished
      description: Needs `can_publish`.
      requestBody: { $ref: '#/components/requestBodies/Published' }
      responses:
        '204': { description: Updated. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/courses/{courseId}/move:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    post:
      tags: [admin-courses]
      summary: Move course
      operationId: adminMoveCourse
      description: Swaps order with the neighbour in the same track; no-op at the edge. Admin only.
      requestBody: { $ref: '#/components/requestBodies/Move' }
      responses:
        '204': { description: Moved. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/courses/{courseId}/cover:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [admin-courses]
      summary: Upload course cover
      operationId: adminUploadCourseCover
      description: Needs `can_edit`.
      requestBody: { $ref: '#/components/requestBodies/CoverUpload' }
      responses:
        '204': { description: Cover replaced. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
    delete:
      tags: [admin-courses]
      summary: Delete course cover
      operationId: adminDeleteCourseCover
      description: Falls back to the generated SVG.
      responses:
        '204': { description: Cover removed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/courses/{courseId}/authors:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [admin-courses]
      summary: List course authors
      operationId: adminListCourseAuthors
      responses:
        '200':
          description: Owner and co-authors.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseAuthors' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [admin-courses]
      summary: Add co-author
      operationId: adminAddCourseAuthor
      description: |
        Needs `can_manage_authors`. The user must have role `author` or `admin`;
        otherwise `422` with `fields.email` = `not_found` or `not_author`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '201':
          description: Added.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorRef' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`already_author` — the user is the owner or a co-author.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/courses/{courseId}/authors/{userId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/UserId'
    delete:
      tags: [admin-courses]
      summary: Remove co-author
      operationId: adminRemoveCourseAuthor
      description: Needs `can_manage_authors`; a co-author may remove themselves.
      responses:
        '204': { description: Removed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/courses/{courseId}/owner:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    put:
      tags: [admin-courses]
      summary: Transfer course ownership
      operationId: adminSetCourseOwner
      description: |
        Admin only. The new owner must have role `author` or `admin` (`422` with
        `fields.user_id`); they stop being a co-author and the previous owner becomes one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: integer }
      responses:
        '200':
          description: Authors after the transfer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseAuthors' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /admin/content/preview:
    post:
      tags: [admin-content]
      summary: Preview content
      operationId: adminPreviewContent
      description: Renders lesson/task content exactly as students will see it. Available to authors.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format, content]
              properties:
                format: { $ref: '#/components/schemas/ContentFormat' }
                content: { type: string }
      responses:
        '200':
          description: Rendered HTML.
          content:
            application/json:
              schema:
                type: object
                required: [html]
                properties:
                  html: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ───────────────────────────── admin: import ─────────────────────────────
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `POST /auth/login`.
  parameters:
    CourseSlug:
      name: courseSlug
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Slug' }
...
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: integer }
...
  requestBodies:
    Published:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [published]
            properties:
              published: { type: boolean }
    Move:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [direction]
            properties:
              direction: { type: string, enum: [up, down] }
    CoverUpload:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [file]
            properties:
              file:
                type: string
                format: binary
                description: PNG, JPEG, WebP or SVG, max 4 MiB.
...
  responses:
    Unauthorized:
      description: '`unauthorized` — no or expired session.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`forbidden` or `csrf_rejected`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: '`not_found`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: '`validation_failed` with per-field codes in `details.fields`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SlugTaken:
      description: '`slug_taken`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PayloadTooLarge:
      description: '`payload_too_large`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnsupportedMedia:
      description: '`unsupported_media_type`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...
    CourseAuthors:
      type: object
      required: [owner, coauthors]
      properties:
        owner:
          oneOf: [{ $ref: '#/components/schemas/AuthorRef' }, { type: 'null' }]
          description: Null for platform (seed) courses, which only admins manage.
        coauthors:
          type: array
          items: { $ref: '#/components/schemas/AuthorRef' }
    AdminCourseInput:
      type: object
      required: [slug, title, track, difficulty]
      properties:
        slug: { $ref: '#/components/schemas/Slug' }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000 }
        track: { type: string, description: 'Existing specialization slug; `gym` is admin only.' }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        category: { type: string, maxLength: 40, description: Empty → derived. }
        label: { $ref: '#/components/schemas/CourseLabel' }
        accent: { type: string, maxLength: 40, description: Cover gradient key; empty → by category. }
        tags:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 40 }
        est_minutes: { type: integer, minimum: 0, maximum: 100000, description: 0 → lessons × 10. }
        cover_url: { type: string, format: uri, description: 'External http(s) cover. Omitted → cover unchanged; remove it via `DELETE …/cover`.' }
        published: { type: boolean, description: 'Omitted → unchanged (draft on create). Changing it needs `can_publish`.' }
    AdminCourse:
      type: object
      required: [id, slug, title, description, track, difficulty, category, label, accent, tags, est_minutes, order_num, published, source, owner, access, cover_url, has_custom_cover, cover_preview_url, created_at]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        track: { type: string }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        category: { type: string, description: Stored value; empty when derived. }
        label: { oneOf: [{ $ref: '#/components/schemas/CourseLabel' }, { type: 'null' }], description: Null when derived. }
        accent: { type: string }
        tags:
          type: array
          items: { type: string }
        est_minutes: { type: integer }
        order_num: { type: integer }
        published: { type: boolean }
        source: { type: string, enum: [seed, admin] }
        owner:
          oneOf: [{ $ref: '#/components/schemas/AuthorRef' }, { type: 'null' }]
        access: { $ref: '#/components/schemas/CourseAccess' }
        cover_url: { type: [string, 'null'], description: External cover URL. }
        has_custom_cover: { type: boolean, description: Uploaded or external cover is set. }
        cover_preview_url: { type: string }
        created_at: { type: string, format: date-time }
    AdminCourseRow:
      allOf:
        - $ref: '#/components/schemas/AdminCourse'
        - type: object
          required: [lessons_count, labs_count]
          properties:
            lessons_count: { type: integer }
            labs_count: { type: integer }
    AdminCourseDetail:
      type: object
      required: [course, lessons]
      properties:
        course: { $ref: '#/components/schemas/AdminCourse' }
        lessons:
          type: array
          items: { $ref: '#/components/schemas/AdminLessonRow' }
...